requirements.txt
aiohttp
aria2p
async_generator
beautifulsoup4>=4.8.1
cowpy
dnspython
emoji
gitpython
google-api-python-client
google-auth-oauthlib
googletrans>=2.4.0
google_images_download>=2.7.1
gTTS>=2.1.1
gTTS-token>=1.1.3
hachoir
httplib2
humanize
heroku3
lyricsgenius
oauth2client
Pillow>=6.0.0
psycopg2-binary
pybase64>=0.4.0
pylast
pySmartDL
python-barcode
python-dotenv
youtube-dl
pytz
qrcode
requests>=2.18.4
speedtest-cli>=2.0.2
sqlalchemy>=1.2
selenium>=3.141.0
telethon>=1.10.6
telethon-session-sqlalchemy>=0.2.6
urbandict>=0.5
wikipedia>=1.4.0
coffeehouse
